Progress
+ WorldSelectScene added
+ WorldCreateScene added
+ GameScene added
+ GameObject added
+ Game for managing GameObjects added
+ LineEdit added

+++ source/gui/canvas.cpp
@@ -12,12 +12,12 @@ void Canvas::update()
 	Vector2 size;
 	if(XWindow::getSize().x < XWindow::getSize().y) {
 		// Fit width and use inverse aspect ratio
-		size.x = min(CANVAS_WIDTH, XWindow::getSize().x);
+		size.x = (float)min(CANVAS_WIDTH, XWindow::getSize().x);
 		size.y = size.x*CANVAS_HEIGHT/CANVAS_WIDTH; 
 	}
 	else {
 		// Fit height and use aspect ratio
-		size.y = min(CANVAS_HEIGHT, XWindow::getSize().y);
+		size.y = (float)min(CANVAS_HEIGHT, XWindow::getSize().y);
 		size.x = size.y*CANVAS_WIDTH/CANVAS_HEIGHT; 
 	}
 	m_rect.size = size/Vector2(XWindow::getSize());
